--- a/src/components/AiResults/styled.ts
+++ b/src/components/AiResults/styled.ts
@@ -11,3 +11,9 @@ export const StyledFlex = styled(Flex)`
     min-height: ${({ theme }) => theme.helpers.remCalc(150)};
   }
 `;
+
+export const StyledAiResults = styled(Flex)`
+  button[role='button'] {
+    display: none;
+  }
+`;