## What?
- Remove the initial AI call which improves performance dramatically for the first page load
- Improve prompt for additional keywords instruction
- Update tooltips, placeholders, copy
- Hide useless controls
- Add 1 min cache to BigCommerce API calls
- Fix Word limit custom use
- Remove additional attributes input

## Why?
UI and tech enhacements

## Testing / Proof
Before:
2s app load time
6s SSR time
![image](https://github.com/bigcommerce/ai-app-foundation/assets/10532978/ff248e7f-8449-4b3d-b836-e865f3fc70cf)

After:
2s app load time
3s SSR time

cached iteration:
less than a 1s for both
![image](https://github.com/bigcommerce/ai-app-foundation/assets/10532978/ff3e8582-b68d-4ed6-8044-5ad865602bc7)


@bigcommerce/team-data